requirements.txt
google-cloud-language==2.2.2
mail-parser==3.15.0
google-cloud-storage==1.38.0